Fixed tab events
The issue was that the player JS was not being reloaded with subsequent
ajax requests. So I moved the coffeescript for it into a function, that
is called both the first time the player loads, and on "refreshPlayer"
events triggers on #contents.

An inline call to trigger refreshPlayer in player.slim does the job.